Usable version of augment+. Fix bug: limit args is wrong.

+++ cli.py
@@ -243,7 +243,7 @@ def main():
                         help="How many data is to be annotated")
     parser.add_argument("--conver_point", type=int, default=0,
                         help="After conver_point, clients with unlabeled data will be involved.")
-    parser.add_argument("--limit", type=int, default=0,
+    parser.add_argument("--limit", type=float, default=0,
                         help="logits < limit will be dropped")
 
     args = parser.parse_args()
